Generates RHS of DiffEqs
Got the right hand side of things done pretty well. Should add some
coefficients, etc—but these are the overall forms.

Given a number n as an argument, generate all the RHS’s where the
number of y’s and the number of primes add up to that value. Pretty
simple.