Added integration connections to layout
Integration connections are now explicitly in the layout output, which
are now handled in fanout. Since they are explicit in layout we no
longer need to worry about them as a special case in fanout, and so
that part of the code has been removed (and finding fans is now simple:
anything with more than one output).